Revert "Merge branch 'master' into mini-atm"
This reverts commit c15ff11fe6cd3f7377f772dfe98d860a101a1bcb, reversing
changes made to f7ecc0629ae4903c7affd4d9bbb0b0fba0191d26.

+++ src/computer.cpp
@@ -12,7 +12,6 @@
 #include <memory>
 #include <utility>
 
-#include "avatar.h"
 #include "coordinate_conversions.h"
 #include "debug.h"
 #include "effect.h"
@@ -428,7 +427,7 @@ void computer::activate_function( computer_action action )
                                     item sewage( "sewage", calendar::turn );
                                     auto candidates = g->m.i_at( x1, y1 );
                                     for( auto &candidate : candidates ) {
-                                        int capa = candidate.get_remaining_capacity_for_liquid( sewage );
+                                        long capa = candidate.get_remaining_capacity_for_liquid( sewage );
                                         if( capa <= 0 ) {
                                             continue;
                                         }
