+++ src/item_stack.cpp
@@ -84,7 +84,7 @@ units::volume item_stack::stored_volume() const
     return ret;
 }
 
-int item_stack::amount_can_fit( const item &it ) const
+long item_stack::amount_can_fit( const item &it ) const
 {
     // Without stacking charges, would we violate the count limit?
     const bool violates_count = size() >= static_cast<size_t>( count_limit() );
@@ -94,7 +94,7 @@ int item_stack::amount_can_fit( const item &it ) const
         return 0l;
     }
     // Call max because a tile may have been overfilled to begin with (e.g. #14115)
-    const int ret = std::max( 0, it.charges_per_volume( free_volume() ) );
+    const long ret = std::max( 0l, it.charges_per_volume( free_volume() ) );
     return it.count_by_charges() ? std::min( ret, it.charges ) : ret;
 }
